refactor: 메인화면-이동전 컴포넌트 분리 #5
