+++ frontend/src/pages/MainPage/components/RouteSection/RouteItem/index.tsx
@@ -1,0 +1,23 @@
+import * as S from './RouteItem.styled';
+
+import Icon, { IconType } from '@/components/Icon';
+
+interface RouteItemProps {
+  icon: IconType;
+  location: string;
+  value: string;
+}
+
+const RouteItem = ({ icon, location, value }: RouteItemProps) => {
+  return (
+    <S.RouteItemLayout>
+      <S.RouteTitleWrapper>
+        <Icon icon={icon} />
+        <S.Location>{location}</S.Location>
+      </S.RouteTitleWrapper>
+      <span>{value || '주소 입력'}</span>
+    </S.RouteItemLayout>
+  );
+};
+
+export default RouteItem;
