+++ frontend/src/pages/MainPage/components/RouteSection/RouteSection.styled.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const RouteSectionLayout = styled.section`
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+`;
+
+export const IconWrapper = styled.div`
+  position: absolute;
+  right: 1.6rem;
+  top: 50%;
+  transform: translateY(-50%);
+`;
